Add parameter alpha to addCurve, move set/getAlpha to AlphaMixIn
AlphaMixIn applies to Image, Scatter and Curve
