+++ silx/gui/plot/backends/BackendBase.py
@@ -85,7 +85,7 @@ class BackendBase(object):
                  color, symbol, linewidth, linestyle,
                  yaxis,
                  xerror, yerror, z, selectable,
-                 fill):
+                 fill, alpha):
         """Add a 1D curve given by x an y to the graph.
 
         :param numpy.ndarray x: The data corresponding to the x axis
@@ -122,6 +122,7 @@ class BackendBase(object):
         :param int z: Layer on which to draw the cuve
         :param bool selectable: indicate if the curve can be selected
         :param bool fill: True to fill the curve, False otherwise
+        :param float alpha: Curve opacity, as a float in [0., 1.]
         :returns: The handle used by the backend to univocally access the curve
         """
         return legend
